Revert "Rename NotifyService folder"
This reverts commit 9ae69d03261a3fec005d9298073993b2c0f7fdb1.
