+++ src/SFA.DAS.Notifications.Infrastructure/NotifyEmailService/GovNotifyPayload.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace SFA.DAS.Notifications.Infrastructure.GovNotifyService
+namespace SFA.DAS.Notifications.Infrastructure.NotifyEmailService
 {
     public class GovNotifyPayload
     {
