+++ src/SFA.DAS.Notifications.Infrastructure/NotifyEmailService/JwtTokenUtility.cs
@@ -2,7 +2,7 @@
 using JWT;
 using Newtonsoft.Json;
 
-namespace SFA.DAS.Notifications.Infrastructure.GovNotifyService
+namespace SFA.DAS.Notifications.Infrastructure.NotifyEmailService
 {
     public class JwtTokenUtility
     {
