+++ src/SFA.DAS.Notifications.Infrastructure/SFA.DAS.Notifications.Infrastructure.csproj
@@ -146,14 +146,14 @@
     <Compile Include="ExecutionPolicies\PolicyNameAttribute.cs" />
     <Compile Include="ExecutionPolicies\RequiredPolicyAttribute.cs" />
     <Compile Include="ExecutionPolicies\SendEmailExecutionPolicy.cs" />
-    <Compile Include="GovNotifyService\GovNotifyPayload.cs" />
-    <Compile Include="GovNotifyService\JwtTokenUtility.cs" />
-    <Compile Include="GovNotifyService\NotifyEmailService.cs" />
-    <Compile Include="GovNotifyService\NotifyHttpClientWrapper.cs" />
-    <Compile Include="GovNotifyService\NotifyMessage.cs" />
-    <Compile Include="GovNotifyService\NotifySmsService.cs" />
-    <Compile Include="GovNotifyService\UnixDate.cs" />
     <Compile Include="LocalEmailService\LocalEmailService.cs" />
+    <Compile Include="NotifyEmailService\GovNotifyPayload.cs" />
+    <Compile Include="NotifyEmailService\JwtTokenUtility.cs" />
+    <Compile Include="NotifyEmailService\NotifyHttpClientWrapper.cs" />
+    <Compile Include="NotifyEmailService\NotifyEmailService.cs" />
+    <Compile Include="NotifyEmailService\NotifyMessage.cs" />
+    <Compile Include="NotifyEmailService\NotifySmsService.cs" />
+    <Compile Include="NotifyEmailService\UnixDate.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SendGridSmtpEmailService\SendGridSmtpEmailService.cs" />
   </ItemGroup>
